requirements.txt
requests
beautifulsoup4
newspaper3k
nltk
PyMuPDF
google-generativeai
python-dotenv
reportlab